Moving domain parsing from deployable artifact to a factory class
