+++ src/main/java/org/mule/tools/apikit/model/MuleDomain.java
@@ -6,31 +6,20 @@
  */
 package org.mule.tools.apikit.model;
 
-import java.io.File;
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-import java.util.jar.JarFile;
 
-import org.apache.commons.io.IOUtils;
 import org.jdom2.Document;
-import org.jdom2.JDOMException;
 import org.jdom2.input.SAXBuilder;
-import org.mule.runtime.api.deployment.meta.MuleDomainModel;
-import org.mule.runtime.api.deployment.persistence.MuleDomainModelJsonSerializer;
 import org.mule.tools.apikit.input.parsers.HttpListenerConfigParser;
 
 public class MuleDomain implements NamedContent, WithConfigs {
 
-  private static final String MULE_ARTIFACT_LOCATION_IN_JAR = "META-INF/mule-artifact/mule-artifact.json";
-  private static final String MULE_DOMAIN_DEFAULT_CONFIG_FILE_NAME = "mule-domain-config.xml";
-
   private List<HttpListenerConfig> configurations;
 
-  private MuleDomain(List<HttpListenerConfig> configurations) {
+  MuleDomain(List<HttpListenerConfig> configurations) {
     this.configurations = configurations;
   }
 
@@ -52,36 +41,6 @@ public class MuleDomain implements NamedContent, WithConfigs {
     Document contentAsDocument = new SAXBuilder().build(content);
     List<HttpListenerConfig> httpListenerConfigs = new HttpListenerConfigParser().parse(contentAsDocument);
     return new MuleDomain(httpListenerConfigs);
-  }
-
-  public static MuleDomain fromDeployableArtifact(File artifact) throws Exception {
-    JarFile jarArtifact = new JarFile(artifact);
-    InputStream muleArtifacts = jarArtifact.getInputStream(jarArtifact.getEntry(MULE_ARTIFACT_LOCATION_IN_JAR));
-    MuleDomainModelJsonSerializer serializer = new MuleDomainModelJsonSerializer();
-    MuleDomainModel domainModel = serializer.deserialize(IOUtils.toString(muleArtifacts));
-
-    Set<String> configs = domainModel.getConfigs();
-
-    if (configs.isEmpty()) {
-      List<HttpListenerConfig> httpListenerConfigs = new ArrayList<>();
-      parseHttpListenerConfigsFromConfigFile(jarArtifact, MULE_DOMAIN_DEFAULT_CONFIG_FILE_NAME, httpListenerConfigs);
-      return new MuleDomain(httpListenerConfigs);
-    } else {
-      List<HttpListenerConfig> httpListenerConfigs = new ArrayList<>();
-      for (String config : configs) {
-        parseHttpListenerConfigsFromConfigFile(jarArtifact, config, httpListenerConfigs);
-      }
-      return new MuleDomain(httpListenerConfigs);
-    }
-  }
-
-  private static void parseHttpListenerConfigsFromConfigFile(JarFile artifact, String configFile,
-                                                             List<HttpListenerConfig> httpListenerConfigs)
-      throws JDOMException, IOException {
-    try (InputStream content = artifact.getInputStream(artifact.getEntry(configFile))) {
-      Document contentAsDocument = new SAXBuilder().build(content);
-      httpListenerConfigs.addAll(new HttpListenerConfigParser().parse(contentAsDocument));
-    }
   }
 
   public static Builder builder() {
